requirements.txt
pyserial
matplotlib
PyQt5
flask
folium